Restrict access to update and delete actions in MediaAdmin view
The update has modified the MediaAdmin/edit.html.twig file to restrict the display of update and delete buttons. These actions are only visible now if the admin.id(object) exists, providing an additional layer of access control. The change improves the management and the security within the media admin view.