Add rounded border styling and modify image display for media boxes
The 'media-box' class in the MediaAdmin list_items.html.twig now includes 'rounded' for a smoother appearance. The image display has also been adjusted, where if the width is defined and present for an object, it will be set as a background image for more uniform presentation. This enhances the visual aesthetics and usability of the Media Admin interface.