Add detailed info to 'fields:added' event in page-admin.js
This update enhances the 'fields:added' event by including additional details, such as the 'id' and 'contentType' in its event data. It also introduces an 'updateLayoutBlocks' function that fetches and updates the layout block settings using Axios, ensuring that the script tags within the fetched HTML are evaluated. These improvements aim to enhance the interactivity and maintainability of the page-admin functionalities.

Signed-off-by: Yorkie Chadwick <y.chadwick@networking.ch>